improvement: use anywidget types and widget.get_state()
