+++ marimo/_plugins/ui/_impl/from_anywidget.py
@@ -9,7 +9,6 @@ from typing import TYPE_CHECKING, Any, Optional
 import marimo._output.data.data as mo_data
 from marimo import _loggers
 from marimo._output.rich_help import mddoc
-from marimo._plugins.core.json_encoder import WebComponentEncoder
 from marimo._plugins.ui._core.ui_element import InitializationArgs, UIElement
 from marimo._plugins.ui._impl.comm import MarimoComm
 from marimo._runtime.functions import Function
@@ -80,8 +79,15 @@ class anywidget(UIElement[T, T]):
         # This gets set to True in super().__init__()
         self._initialized = False
 
-        # Get all the traits of the widget
-        args: T = widget.trait_values()
+        import ipywidgets  # type: ignore
+
+        _remove_buffers = ipywidgets.widgets.widget._remove_buffers  # type: ignore
+
+        # Get state with custom serializers properly applied
+        state = widget.get_state()
+        state_no_buffers, buffer_paths, buffers = _remove_buffers(state)  # type: ignore
+
+        # Remove widget-specific system traits not needed for the frontend
         ignored_traits = [
             "comm",
             "layout",
@@ -104,30 +110,16 @@ class anywidget(UIElement[T, T]):
             "_view_module_version",
             "_view_name",
         ]
-        # Remove ignored traits
-        for trait_name in ignored_traits:
-            args.pop(trait_name, None)
 
-        # Keep only classes that are json serialize-able
-        json_args: T = {}
-        for k, v in args.items():
-            try:
-                # Try to see if it is json-serializable
-                WebComponentEncoder.json_dumps(v)
-                # Just add the plain value, it will be json-serialized later
-                json_args[k] = v
-            except TypeError:
-                pass
-            except ValueError:
-                # Handle circular dependencies
-                pass
+        # Filter out system traits from the serialized state
+        json_args: T = {
+            k: v
+            for k, v in state_no_buffers.items()
+            if k not in ignored_traits
+        }
 
         js: str = widget._esm if hasattr(widget, "_esm") else ""  # type: ignore [unused-ignore]  # noqa: E501
         css: str = widget._css if hasattr(widget, "_css") else ""  # type: ignore [unused-ignore]  # noqa: E501
-        import ipywidgets  # type: ignore
-
-        _remove_buffers = ipywidgets.widgets.widget._remove_buffers  # type: ignore
-        _state, buffer_paths, buffers = _remove_buffers(widget.get_state())  # type: ignore
 
         def on_change(change: T) -> None:
             _put_buffers = ipywidgets.widgets.widget._put_buffers  # type: ignore
@@ -180,6 +172,10 @@ class anywidget(UIElement[T, T]):
             merged = {**self._prev_state, **value}
             self._prev_state = merged
             return merged
+
+        LOGGER.warning(
+            f"Expected anywidget value to be a dict, got {type(value)}"
+        )
         self._prev_state = value
         return value
